requirements.txt
fastapi          
google-genai     
PyPDF2          
python-docx      
python-dotenv    
python-multipart 
uvicorn          